[DashboardBundle] Fix ConfigHelper that requires kuma_analytics_config table to be present on constuct | Q      | A | | ----------- | ----------- | | Bug fix?      | Yes       | | New feature?   | No        | | BC breaks?   | No        | | Deprecations?   | No        | | Deprecations?   | No        | | Fixed tickets   | #3044         |
Note:
It's not considered good practice to call `init` methods in `__construct` methods.
This PR does not fix the underlying problem and introduces a new query every time this service is used to check if the required table exists.
A better solution would be to rework the usage of `\Kunstmaan\DashboardBundle\Helper\Google\Analytics\ConfigHelper` to remove the `init` method from the constructor and enforce the `init` call wherever the service is used. This however would cause BC breaks.